FINALLY moving mockRunLogs to armi.testing
